refactor: merge compat module components and fix line length
